Refactored not to use grad_conv schema and removed the related flyway stuff.
Refactored not to use grad_conv schema and removed the related flyway stuff.
