+++ api/src/main/java/ca/bc/gov/educ/api/dataconversion/config/TraxDbConfig.java
@@ -5,6 +5,7 @@ import com.zaxxer.hikari.HikariDataSource;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.orm.jpa.JpaTransactionManager;
@@ -53,6 +54,7 @@ public class TraxDbConfig {
     @Value("${spring.db-connection.trax.password}")
     private String traxPassword;
 
+    @Primary
     @Bean
     public DataSource traxDataSource() {
         HikariConfig config = new HikariConfig();
@@ -71,6 +73,7 @@ public class TraxDbConfig {
         return new HikariDataSource(config);
     }
 
+    @Primary
     @Bean
     public LocalContainerEntityManagerFactoryBean traxEntityManager() {
         LocalContainerEntityManagerFactoryBean em
@@ -92,6 +95,7 @@ public class TraxDbConfig {
         return em;
     }
 
+    @Primary
     @Bean
     public PlatformTransactionManager traxTransactionManager() {
         JpaTransactionManager transactionManager = new JpaTransactionManager();
