+++ api/src/main/java/ca/bc/gov/educ/api/dataconversion/entity/trax/GraduationCourseKey.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.educ.api.dataconversion.entity.conv;
+package ca.bc.gov.educ.api.dataconversion.entity.trax;
 
 import lombok.Data;
