+++ api/src/main/resources/application.yaml
@@ -24,10 +24,6 @@ spring:
       username: ${BATCH_USERNAME}
       password: ${BATCH_PASSWORD}
       pool-name: conv-batch-cp
-    conv:
-      username: ${CONV_USERNAME}
-      password: ${CONV_PASSWORD}
-      pool-name: conv-data-conv-cp
     course:
       username: ${COURSE_USERNAME}
       password: ${COURSE_PASSWORD}
@@ -62,15 +58,6 @@ spring:
     initialize-schema: NEVER
     job:
       enabled: false
-  flyway:
-    enabled: ${ENABLE_FLYWAY}
-    baseline-on-migrate: false
-    table: FLYWAY_SCHEMA_HISTORY
-    url: ${CONNECTION_STRING}
-    schema: ${CONV_USERNAME}
-    user: ${CONV_USERNAME}
-    password: ${CONV_PASSWORD}
-    location: classpath:db/migration
 
 #Logging properties
 logging:
