--- api/src/main/resources/db/migration/1.0/V1.0.2__DDL-CREATE_SYNONYM-trax_students_load.sql
@@ -1,1 +1,0 @@
-CREATE OR REPLACE EDITIONABLE SYNONYM "TRAX_STUDENTS_LOAD" FOR "TRAX_STUDENTS_LOAD"@"TRAXLINK.WORLD";
