--- api/src/main/resources/db/migration/1.0/V1.0.3__DDL-CREATE_SYNONYM-student_master.sql
@@ -1,1 +1,0 @@
-CREATE OR REPLACE EDITIONABLE SYNONYM "STUDENT_MASTER" FOR "STUDENT_MASTER"@"TRAXLINK.WORLD";
