--- api/src/main/resources/db/migration/1.0/V1.0.4__DDL-CREATE_SYNONYM-stud_xcrse.sql
@@ -1,1 +1,0 @@
-CREATE OR REPLACE EDITIONABLE SYNONYM "STUD_XCRSE" FOR "STUD_XCRSE"@"TRAXLINK.WORLD";
