--- api/src/main/resources/db/migration/1.0/V1.0.6__DDL-CREATE_TABLE-conv_course_restriction.sql
@@ -1,19 +1,0 @@
--- COURSE_RESTRICTION definition
-
-CREATE TABLE "CONV_COURSE_RESTRICTION"
-   (	"COURSE_RESTRICTION_ID" RAW(16) DEFAULT SYS_GUID(), 
-	"MAIN_COURSE" VARCHAR2(7) NOT NULL ENABLE, 
-	"MAIN_COURSE_LEVEL" VARCHAR2(3), 
-	"RESTRICTED_COURSE" VARCHAR2(7) NOT NULL ENABLE, 
-	"RESTRICTED_COURSE_LVL" VARCHAR2(3), 
-	"RESTRICTION_EFFECTIVE_DATE" DATE NOT NULL ENABLE, 
-	"RESTRICTION_EXPIRY_DATE" DATE NOT NULL ENABLE, 
-	"CREATE_USER" VARCHAR2(32) DEFAULT USER NOT NULL ENABLE, 
-	"CREATE_DATE" DATE DEFAULT SYSTIMESTAMP NOT NULL ENABLE, 
-	"UPDATE_USER" VARCHAR2(32) DEFAULT USER NOT NULL ENABLE, 
-	"UPDATE_DATE" DATE DEFAULT SYSTIMESTAMP NOT NULL ENABLE, 
-	 CONSTRAINT "COURSE_RESTRICTION_PK" PRIMARY KEY ("COURSE_RESTRICTION_ID")
-  USING INDEX TABLESPACE "API_GRAD_IDX"  ENABLE
-   ) SEGMENT CREATION IMMEDIATE 
- NOCOMPRESS LOGGING
-  TABLESPACE "API_GRAD_DATA"   NO INMEMORY ;
