--- api/src/main/resources/db/migration/1.0/V1.0.7__DDL-CREATE_TABLE-conv_graduation_student_record.sql
@@ -1,23 +1,0 @@
--- COURSE_RESTRICTION definition
-
-CREATE TABLE "CONV_GRADUATION_STUDENT_RECORD"
-   (    "GRADUATION_STUDENT_RECORD_ID" RAW(16) DEFAULT SYS_GUID(),
-    "GRADUATION_PROGRAM_CODE" VARCHAR2(8 BYTE) NOT NULL ENABLE,
-    "PROGRAM_COMPLETION_DATE" DATE,
-    "GPA" NUMBER(5,4),
-    "HONOURS_STANDING" VARCHAR2(1 BYTE),
-    "RECALCULATE_GRAD_STATUS" VARCHAR2(1 BYTE),
-    "STUDENT_GRAD_DATA" CLOB,
-    "SCHOOL_OF_RECORD" VARCHAR2(9 BYTE),
-    "SCHOOL_AT_GRADUATION" VARCHAR2(9 BYTE),
-    "STUDENT_GRADE" VARCHAR2(3 BYTE),
-    "STUDENT_STATUS_CODE" VARCHAR2(1 BYTE),
-	"CREATE_USER" VARCHAR2(32) DEFAULT USER NOT NULL ENABLE, 
-	"CREATE_DATE" DATE DEFAULT SYSTIMESTAMP NOT NULL ENABLE, 
-	"UPDATE_USER" VARCHAR2(32) DEFAULT USER NOT NULL ENABLE, 
-	"UPDATE_DATE" DATE DEFAULT SYSTIMESTAMP NOT NULL ENABLE, 
-	 CONSTRAINT "CONV_GRADUATION_STUDENT_RECORD_PK" PRIMARY KEY ("GRADUATION_STUDENT_RECORD_ID")
-  USING INDEX TABLESPACE "API_GRAD_IDX"  ENABLE
-   ) SEGMENT CREATION IMMEDIATE 
- NOCOMPRESS LOGGING
-  TABLESPACE "API_GRAD_DATA"   NO INMEMORY ;
