--- api/src/main/resources/db/migration/1.0/V1.0.8__DDL-CREATE_TABLE-conv_student_optional_program.sql
@@ -1,19 +1,0 @@
--- COURSE_RESTRICTION definition
-
-CREATE TABLE "CONV_STUDENT_OPTIONAL_PROGRAM"
-   (    "STUDENT_OPTIONAL_PROGRAM_ID" RAW(16) DEFAULT SYS_GUID(),
-    "OPTIONAL_PROGRAM_ID" RAW(16) NOT NULL ENABLE,
-    "GRADUATION_STUDENT_RECORD_ID" RAW(16) NOT NULL ENABLE,
-    "COMPLETION_DATE" DATE,
-    "PROGRAM_NOTE" CLOB,
-	"CREATE_USER" VARCHAR2(32) DEFAULT USER NOT NULL ENABLE, 
-	"CREATE_DATE" DATE DEFAULT SYSTIMESTAMP NOT NULL ENABLE, 
-	"UPDATE_USER" VARCHAR2(32) DEFAULT USER NOT NULL ENABLE, 
-	"UPDATE_DATE" DATE DEFAULT SYSTIMESTAMP NOT NULL ENABLE, 
-	 CONSTRAINT "CONV_STUDENT_OPTIONAL_PROGRAM_PK" PRIMARY KEY ("STUDENT_OPTIONAL_PROGRAM_ID")
-  USING INDEX TABLESPACE "API_GRAD_IDX"  ENABLE,
-     CONSTRAINT "FK_CONV_GRADUATION_STUDENT_RECORD" FOREIGN KEY ("GRADUATION_STUDENT_RECORD_ID")
-        REFERENCES "CONV_GRADUATION_STUDENT_RECORD" ("GRADUATION_STUDENT_RECORD_ID") ENABLE
-   ) SEGMENT CREATION IMMEDIATE 
- NOCOMPRESS LOGGING
-  TABLESPACE "API_GRAD_DATA"   NO INMEMORY ;
