--- api/src/main/resources/db/migration/1.1/V1.1.1__DDL-DROP_TABLE-spring_batch.sql
@@ -1,11 +1,0 @@
--- Spring Batch Tables definition
-DROP TABLE BATCH_STEP_EXECUTION_CONTEXT;
-DROP TABLE BATCH_STEP_EXECUTION;
-DROP TABLE BATCH_JOB_EXECUTION_CONTEXT;
-DROP TABLE BATCH_JOB_EXECUTION_PARAMS;
-DROP TABLE BATCH_JOB_EXECUTION;
-DROP TABLE BATCH_JOB_INSTANCE;
-
-DROP SEQUENCE BATCH_STEP_EXECUTION_SEQ;
-DROP SEQUENCE BATCH_JOB_EXECUTION_SEQ;
-DROP SEQUENCE BATCH_JOB_SEQ;
