--- api/src/main/resources/db/migration/1.1/V1.1.2__DDL-DROP_SYNONYM-stud_xcrse.sql
@@ -1,1 +1,0 @@
-DROP SYNONYM "STUD_XCRSE";
