--- api/src/main/resources/db/migration/1.1/V1.1.3__DDL-DROP_TABLE-conv_student_optional_program.sql
@@ -1,2 +1,0 @@
-DROP TABLE "CONV_STUDENT_OPTIONAL_PROGRAM";
-
