--- api/src/main/resources/db/migration/1.1/V1.1.4__DDL-CREATE_SYNONYM-tab_grad_crse.sql
@@ -1,1 +1,0 @@
-CREATE OR REPLACE EDITIONABLE SYNONYM "TAB_GRAD_CRSE" FOR "TAB_GRAD_CRSE"@"TRAXLINK.WORLD";
