--- api/src/main/resources/db/migration/1.1/V1.1.5__DDL-DROP_TABLE-conv_course_restriction.sql
@@ -1,1 +1,0 @@
-DROP TABLE "CONV_COURSE_RESTRICTION";
