--- api/src/main/resources/db/migration/1.1/V1.1.6__DDL-DROP_TABLE-conv_graduation_student_record.sql
@@ -1,1 +1,0 @@
-DROP TABLE "CONV_GRADUATION_STUDENT_RECORD";
