+++ api/src/test/java/ca/bc/gov/educ/api/dataconversion/config/TraxDbConfig.java
@@ -5,6 +5,7 @@ import com.zaxxer.hikari.HikariDataSource;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.context.annotation.Profile;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.orm.jpa.JpaTransactionManager;
@@ -47,6 +48,7 @@ public class TraxDbConfig {
     @Value("${spring.datasource.url}")
     private String jdbcUrl;
 
+    @Primary
     @Bean
     public DataSource traxDataSource() {
         HikariConfig config = new HikariConfig();
@@ -64,6 +66,7 @@ public class TraxDbConfig {
         return new HikariDataSource(config);
     }
 
+    @Primary
     @Bean
     public LocalContainerEntityManagerFactoryBean traxEntityManager() {
         LocalContainerEntityManagerFactoryBean em
@@ -85,6 +88,7 @@ public class TraxDbConfig {
         return em;
     }
 
+    @Primary
     @Bean
     public PlatformTransactionManager traxTransactionManager() {
         JpaTransactionManager transactionManager = new JpaTransactionManager();
