+++ api/src/test/java/ca/bc/gov/educ/api/dataconversion/service/DataConversionServiceTest.java
@@ -1,7 +1,7 @@
 package ca.bc.gov.educ.api.dataconversion.service;
 
 import ca.bc.gov.educ.api.dataconversion.model.*;
-import ca.bc.gov.educ.api.dataconversion.repository.conv.TraxStudentsLoadRepository;
+import ca.bc.gov.educ.api.dataconversion.repository.trax.TraxStudentsLoadRepository;
 import ca.bc.gov.educ.api.dataconversion.service.conv.DataConversionService;
 import ca.bc.gov.educ.api.dataconversion.util.RestUtils;
 import org.junit.After;
